# docs/swagger.yaml
basePath: /v1
definitions:
  main.CreatePostPayload:
    properties:
      message:
        maxLength: 1000
        minLength: 20
        type: string
      tags:
        items:
          type: string
        type: array
      title:
        maxLength: 100
        minLength: 8
        type: string
    required:
    - message
    - title
    type: object
  main.CreateUserTokenPayload:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  main.RegisterUserPayload:
    properties:
      email:
        maxLength: 255
        type: string
      password:
        maxLength: 72
        minLength: 8
        type: string
      username:
        maxLength: 100
        type: string
    required:
    - email
    - password
    - username
    type: object
  main.UpdatePostPayload:
    properties:
      message:
        maxLength: 100
        type: string
      tags:
        items:
          type: string
        type: array
      title:
        maxLength: 100
        type: string
    type: object
  main.UserIdPayload:
    properties:
      user_id:
        type: string
    type: object
  main.UserTokenPayload:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      role:
        $ref: '#/definitions/store.Role'
      role_id:
        type: integer
      token:
        type: string
      username:
        type: string
    type: object
  store.Comment:
    properties:
      content:
        type: string
      created_at:
        type: string
      id:
        type: integer
      post_id:
        type: integer
      user:
        $ref: '#/definitions/store.User'
      user_id:
        type: integer
    type: object
  store.Post:
    properties:
      comments:
        items:
          $ref: '#/definitions/store.Comment'
        type: array
      content:
        type: string
      created_at:
        type: string
      id:
        type: integer
      tags:
        items:
          type: string
        type: array
      title:
        type: string
      updated_at:
        type: string
      user:
        $ref: '#/definitions/store.User'
      user_id:
        type: integer
      version:
        type: integer
    type: object
  store.Role:
    properties:
      description:
        type: string
      id:
        type: integer
      level:
        type: integer
      name:
        type: string
    type: object
  store.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      role:
        $ref: '#/definitions/store.Role'
      role_id:
        type: integer
      username:
        type: string
    type: object
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is an API for a social network.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Social Kib
paths:
  /authentication/token:
    post:
      consumes:
      - application/json
      description: Create a token for a user
      parameters:
      - description: User credentials
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/main.CreateUserTokenPayload'
      produces:
      - application/json
      responses:
        "200":
          description: Token
          schema:
            type: string
        "400":
          description: Bad Request
          schema: {}
        "404":
          description: Not Found
          schema: {}
      summary: Create a token
      tags:
      - auth
  /authentication/user:
    post:
      consumes:
      - application/json
      description: YOu can create an account
      parameters:
      - description: FollowerId
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/main.RegisterUserPayload'
      produces:
      - application/json
      responses:
        "200":
          description: User registered
          schema:
            $ref: '#/definitions/main.UserTokenPayload'
        "400":
          description: Bad Request
          schema: {}
        "404":
          description: Not Found
          schema: {}
      summary: Create a new account
      tags:
      - auth
  /feed:
    get:
      consumes:
      - application/json
      description: Get a user
      parameters:
      - description: Limit search of query
        in: query
        name: limit
        type: integer
      - description: Offset for each search
        in: query
        name: offset
        type: integer
      - description: Sort asc or desc
        in: query
        name: sort
        type: string
      - description: search title or content
        in: query
        name: search
        type: string
      - description: search since the post
        in: query
        name: since
        type: string
      - collectionFormat: csv
        description: search post from tags
        in: query
        items:
          type: string
        name: tags
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/store.User'
        "400":
          description: Bad Request
          schema: {}
        "404":
          description: Not Found
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Get the user feed
      tags:
      - users
  /posts:
    post:
      consumes:
      - application/json
      description: Create post
      parameters:
      - description: Create post payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/main.CreatePostPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/store.Post'
        "400":
          description: Bad Request
          schema: {}
        "404":
          description: Not Found
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Create post
      tags:
      - posts
  /posts/{id}:
    get:
      consumes:
      - application/json
      description: GEt post by id
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/store.Post'
        "400":
          description: Bad Request
          schema: {}
        "404":
          description: Not Found
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Get a post by the id
      tags:
      - posts
    patch:
      consumes:
      - application/json
      description: Update post by id
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update post Payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/main.UpdatePostPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/store.Post'
        "400":
          description: Bad Request
          schema: {}
        "404":
          description: Not Found
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Update a post by the id
      tags:
      - posts
  /posts/{postId}:
    delete:
      consumes:
      - application/json
      description: Delete post by id
      parameters:
      - description: Post ID
        in: path
        name: postId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/store.Post'
        "400":
          description: Bad Request
          schema: {}
        "404":
          description: Not Found
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Delete a post by the id
      tags:
      - posts
  /users/{id}:
    get:
      consumes:
      - application/json
      description: Fetches a user profile by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/store.User'
        "400":
          description: Bad Request
          schema: {}
        "404":
          description: Not Found
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Fetches a user profile
      tags:
      - users
  /users/activate/{token}:
    put:
      consumes:
      - application/json
      description: The api will confirm the token send previouse on the email
      parameters:
      - description: Token
        in: path
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/store.User'
        "400":
          description: Bad Request
          schema: {}
        "404":
          description: Not Found
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Activate user by activating the token
      tags:
      - users
  /users/follow/{id}:
    post:
      consumes:
      - application/json
      description: YOu can follow a user and keep update with their posts
      parameters:
      - description: FollowerId
        in: body
        name: id
        required: true
        schema:
          $ref: '#/definitions/main.UserIdPayload'
      - description: UserID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/store.Post'
        "400":
          description: Bad Request
          schema: {}
        "404":
          description: Not Found
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Follow a new user
      tags:
      - users
  /users/unfollow/{id}:
    delete:
      consumes:
      - application/json
      description: YOu can follow a user and keep update with their posts
      parameters:
      - description: FollowerId
        in: body
        name: id
        required: true
        schema:
          $ref: '#/definitions/main.UserIdPayload'
      - description: UserID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/store.Post'
        "400":
          description: Bad Request
          schema: {}
        "404":
          description: Not Found
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Follow a new user
      tags:
      - users
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"