# store/comments.go
package store

import (
	"context"
	"database/sql"
	"fmt"
)

type CommentStorage struct {
	db *sql.DB
}
type Comment struct {
	ID        int64  `json:"id"`
	PostID    int64  `json:"post_id"`
	UserID    int64  `json:"user_id"`
	Content   string `json:"content"`
	CreatedAt string `json:"created_at"`
	User      User   `json:"user"`
}

func (s *CommentStorage) Create(ctx context.Context, comment *Comment) error {
	query := `
		INSERT INTO comments (post_id, user_id, content)
		VALUES ($1,$2,$3) 
		RETURNING id, post_id, user_id, content
	`
	err := s.db.QueryRowContext(ctx, query, comment.PostID, comment.UserID, comment.Content).Scan(&comment.ID, &comment.PostID, &comment.UserID, &comment.Content)
	if err != nil {
		return err
	}
	return nil
}
func (s *CommentStorage) GetCommentsByUserId(ctx context.Context, postID int64) ([]Comment, error) {
	query := `
		SELECT c.id, c.post_id, c.user_id, c.content, c.created_at, users.username, users.id FROM comments c
		JOIN users on users.id = c.user_id
		WHERE c.post_id =$1
		ORDER BY c.created_at DESC;
	`

	rows, err := s.db.QueryContext(ctx, query, postID)
	if err != nil {
		return nil, err
	}

	comments := []Comment{}
	for rows.Next() {
		var c Comment
		c.User = User{}
		err := rows.Scan(&c.ID, &c.PostID, &c.UserID, &c.Content, &c.CreatedAt, &c.User.Username, &c.User.ID)
		if err != nil {
			fmt.Println("----------------COMMMMMMMMMMMMMMEEEEEEEEEEEEEEENTS", &rows)
			return nil, err
		}
		comments = append(comments, c)

	}
	return comments, nil
}